# docs/swagger.yaml
basePath: /
definitions:
...
  groups.CreateGroupRequest:
    properties:
      description:
        type: object
      name:
        type: object
    type: object
  groups.ErrorResponse:
    properties:
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid group name
        type: string
    type: object
  groups.GroupListResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/groups.GroupResponse'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  groups.GroupResponse:
    properties:
      description:
        type: object
      id:
        example: 1
        type: integer
      name:
        type: object
      public_id:
        example: grp-abc123
        type: string
    type: object
  groups.GroupRoleResponse:
    properties:
      group_public_id:
        example: grp-abc123
        type: string
      role_id:
        example: 1
        type: integer
      role_name:
        example: admin
        type: string
    type: object
...
  groups.SuccessResponse:
    properties:
      message:
        example: Group deleted successfully
        type: string
    type: object
  groups.UpdateGroupRequest:
    properties:
      description:
        type: object
      name:
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@knowledgecenter.io
    name: API Support
  description: Knowledge Center REST API Server
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: Knowledge Center API
  version: "1.0"
paths:
...
  /groups:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of groups
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.GroupListResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Creates a new group with JSONB name and description
      parameters:
      - description: Group data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/groups.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/groups.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new group
      tags:
      - groups
  /groups/{publicId}:
    delete:
      consumes:
      - application/json
      description: Deletes a group by public ID
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete group
      tags:
      - groups
    get:
      consumes:
      - application/json
      description: Retrieves a group by its public ID
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.GroupResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get group by public ID
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Updates an existing group
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      - description: Group data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/groups.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update group
      tags:
      - groups
  /groups/{publicId}/roles:
    get:
      consumes:
      - application/json
      description: Retrieves all roles assigned to a group
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/groups.GroupRoleResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get group roles
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Assigns roles to a group (replaces existing roles)
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      - description: Role IDs to assign
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/groups.AssignRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign roles to group
      tags:
      - groups
  /groups/{publicId}/roles/{roleId}:
    delete:
      consumes:
      - application/json
      description: Removes a role from a group
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove role from group
      tags:
      - groups
  /groups/{publicId}/users:
    get:
      consumes:
      - application/json
      description: Retrieves all users in a group
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get group users
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Assigns users to a group (replaces existing users)
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      - description: User IDs to assign
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/groups.AssignUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign users to group
      tags:
      - groups
  /groups/{publicId}/users/{userId}:
    delete:
      consumes:
      - application/json
      description: Removes a user from a group
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove user from group
      tags:
      - groups
  /groups/batch:
    delete:
      consumes:
      - application/json
      description: Deletes multiple groups in a single transaction
      parameters:
      - description: Batch delete request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/groups.BatchDeleteGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch delete groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Creates multiple groups in a single transaction
      parameters:
      - description: Batch create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/groups.BatchCreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/groups.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch create groups
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Updates multiple groups in a single transaction
      parameters:
      - description: Batch update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/groups.BatchUpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch update groups
      tags:
      - groups
...
  /users/{userId}/groups:
    get:
      consumes:
      - application/json
      description: Retrieves all groups a user belongs to
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/groups.GroupResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user groups
      tags:
      - groups
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"